This commit was manufactured by cvs2svn to create tag 'Ext-5-1-1'.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/tags/Ext-5-1-1@9940 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/testing/eval_tools.sh
@@ -349,8 +349,11 @@ STARTPROG() {
         COMMAND="$COMMAND $PORT_SPEC"
     fi
     echo $COMMAND >> $SNMP_TMPDIR/invoked
-    $COMMAND > $LOG_FILE.stdout 2>&1
-
+    if [ "x$OSTYPE" = "xmsys" ]; then
+      $COMMAND > $LOG_FILE.stdout 2>&1 &
+    else
+      $COMMAND > $LOG_FILE.stdout 2>&1
+    fi
     DELAY
 }
 
@@ -364,11 +367,13 @@ STARTPROGNOSLEEP() {
 #------------------------------------ -o-
 STARTAGENT() {
     SNMPDSTARTED=1
-    COMMAND="snmpd $SNMP_FLAGS -r -U -P $SNMP_SNMPD_PID_FILE -l $SNMP_SNMPD_LOG_FILE $AGENT_FLAGS"
+    COMMAND="snmpd $SNMP_FLAGS -r -U -p $SNMP_SNMPD_PID_FILE -Lf $SNMP_SNMPD_LOG_FILE $AGENT_FLAGS"
     CFG_FILE=$SNMP_CONFIG_FILE
     LOG_FILE=$SNMP_SNMPD_LOG_FILE
     PORT_SPEC="$SNMP_SNMPD_PORT"
-
+    if [ "x$SNMP_TRANSPORT_SPEC" != "x" ]; then
+        PORT_SPEC="$SNMP_TRANSPORT_SPEC:$PORT_SPEC"
+    fi
     STARTPROGNOSLEEP
     WAITFORAGENT "NET-SNMP version"
 }
@@ -376,11 +381,13 @@ STARTAGENT() {
 #------------------------------------ -o-
 STARTTRAPD() {
     TRAPDSTARTED=1
-    COMMAND="snmptrapd -d -u $SNMP_SNMPTRAPD_PID_FILE -o $SNMP_SNMPTRAPD_LOG_FILE"
+    COMMAND="snmptrapd -d -p $SNMP_SNMPTRAPD_PID_FILE -Lf $SNMP_SNMPTRAPD_LOG_FILE"
     CFG_FILE=$SNMPTRAPD_CONFIG_FILE
     LOG_FILE=$SNMP_SNMPTRAPD_LOG_FILE
     PORT_SPEC="$SNMP_SNMPTRAPD_PORT"
-
+    if [ "x$SNMP_TRANSPORT_SPEC" != "x" ]; then
+        PORT_SPEC="$SNMP_TRANSPORT_SPEC:$PORT_SPEC"
+    fi
     STARTPROGNOSLEEP
     WAITFORTRAPD "NET-SNMP version"
 }
@@ -392,7 +399,11 @@ STARTTRAPD() {
 #    master agent and sub agent.
 STOPPROG() {
     if [ -f $1 ]; then
-	COMMAND="kill -TERM `cat $1`"
+        if [ "x$OSTYPE" = "xmsys" ]; then
+          COMMAND="kill.exe `cat $1`"
+        else
+          COMMAND="kill -TERM `cat $1`"
+        fi
 	echo $COMMAND >> $SNMP_TMPDIR/invoked
 
 	DELAY
@@ -412,7 +423,9 @@ STOPPROGNOSLEEP() {
 STOPAGENT() {
     SAVE_RESULTS
     STOPPROGNOSLEEP $SNMP_SNMPD_PID_FILE
-    WAITFORAGENT "shutting down"
+    if [ "x$OSTYPE" != "xmsys" ]; then
+        WAITFORAGENT "shutting down"
+    fi
     if [ $SNMP_VERBOSE -gt 1 ]; then
 	echo "Agent Output:"
 	echo "$seperator [stdout]"
@@ -428,7 +441,9 @@ STOPAGENT() {
 STOPTRAPD() {
     SAVE_RESULTS
     STOPPROGNOSLEEP $SNMP_SNMPTRAPD_PID_FILE
-    WAITFORTRAPD "Stopped"
+    if [ "x$OSTYPE" != "xmsys" ]; then
+        WAITFORTRAPD "Stopped"
+    fi
     if [ $SNMP_VERBOSE -gt 1 ]; then
 	echo "snmptrapd Output:"
 	echo "$seperator [stdout]"
@@ -453,7 +468,11 @@ FINISHED() {
 	ps -e | egrep "^[ ]*$pid" > /dev/null 2>&1
 	if [ $? = 0 ] ; then
 	    SNMP_SAVE_TMPDIR=yes
-	    COMMAND="kill -9 $pid"
+            if [ "x$OSTYPE" = "xmsys" ]; then
+              COMMAND="kill -9 $pid"
+            else
+              COMMAND="kill.exe $pid"
+            fi
 	    echo $COMMAND "($pfile)" >> $SNMP_TMPDIR/invoked
 	    $COMMAND > /dev/null 2>&1
 	    return_value=1
